Changed database. Fixed types.

+++ TeamManagementApp/Models/Filters/ProjectFilter.cs
@@ -14,10 +14,10 @@ namespace TeamManagementApp.Models.Filters
         public Selection ProjectSelection { get; set; } = Selection.Any;
         public string Project { get; set; } = string.Empty;
         public Selection AssigneeSelection {  get; set; } = Selection.Any;
-        public List<int> Assignees { get; set; } = [];
+        public List<long> Assignees { get; set; } = [];
         public Selection StatusSelection { get; set; } = Selection.Any;
-        public List<int> Statuses { get; set; } = [];
+        public List<long> Statuses { get; set; } = [];
         public Selection TypeSelection { get; set; } = Selection.Any;
-        public List<int> Types { get; set; } = [];
+        public List<long> Types { get; set; } = [];
     }
 }
