+++ TeamManagementApp/Models/Member.cs
@@ -8,10 +8,10 @@ namespace TeamManagementApp.Models
 {
     public class Member
     {
-        public int MemberId { get; set; }
+        public long MemberId { get; set; }
         public string Name { get; set; } = string.Empty;
         public bool Active { get; set; }
-        public int ClassId { get; set; }
-        public int? UserId { get; set; }
+        public long ClassId { get; set; }
+        public long? UserId { get; set; }
     }
 }
